+++ packages/eslint-plugin/configs/recommended.js
@@ -50,6 +50,9 @@ if ( isPackageInstalled( 'typescript' ) ) {
 				'jsdoc/require-returns-type': 'off',
 				// handled by TS itself
 				'no-unused-vars': 'off',
+				// no-shadow doesn't work correctly in TS, so let's use a TS-dedicated version instead.
+				'no-shadow': 'off',
+				'@typescript-eslint/no-shadow': 'error',
 			},
 		},
 	];